requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via -r requirements.in
aiosignal==1.4.0
    # via aiohttp
allure-pytest==2.15.2
    # via -r requirements.in
allure-python-commons==2.15.2
    # via allure-pytest
anyio==4.12.0
    # via
    #   httpx
    #   starlette
arrow==1.4.0
    # via isoduration
attrs==25.4.0
    # via
    #   aiohttp
    #   allure-python-commons
    #   jsonschema
    #   pytest-subtests
    #   referencing
bidict==0.23.1
    # via python-socketio
black==25.12.0
    # via -r requirements.in
blinker==1.9.0
    # via flask
brotli==1.2.0
    # via geventhttpclient
certifi==2025.11.12
    # via
    #   geventhttpclient
    #   httpcore
    #   httpx
    #   requests
chardet==5.2.0
    # via prance
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   black
    #   flask
    #   schemathesis
colorama==0.4.6
    # via
    #   pactman
    #   schemathesis
configargparse==1.7.1
    # via
    #   locust
    #   locust-cloud
contourpy==1.3.3
    # via matplotlib
coverage[toml]==7.13.0
    # via pytest-cov
cycler==0.12.1
    # via matplotlib
execnet==2.1.2
    # via pytest-xdist
faker==38.2.0
    # via -r requirements.in
flask==3.1.2
    # via
    #   flask-cors
    #   flask-login
    #   locust
flask-cors==6.0.1
    # via locust
flask-login==0.6.3
    # via locust
fonttools==4.61.0
    # via matplotlib
fqdn==1.5.1
    # via jsonschema
freezegun==1.5.5
    # via -r requirements.in
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
gevent==25.9.1
    # via
    #   geventhttpclient
    #   locust
    #   locust-cloud
geventhttpclient==2.3.7
    # via locust
graphql-core==3.2.7
    # via hypothesis-graphql
greenlet==3.3.0
    # via gevent
h11==0.16.0
    # via
    #   httpcore
    #   wsproto
harfile==0.4.0
    # via schemathesis
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via schemathesis
hypothesis==6.148.7
    # via
    #   hypothesis-graphql
    #   hypothesis-jsonschema
    #   schemathesis
hypothesis-graphql==0.11.1
    # via schemathesis
hypothesis-jsonschema==0.23.1
    # via schemathesis
idna==3.11
    # via
    #   anyio
    #   httpx
    #   jsonschema
    #   requests
    #   yarl
iniconfig==2.3.0
    # via pytest
isoduration==20.11.0
    # via jsonschema
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   flask
    #   pytest-html
jsonpath-ng==1.7.0
    # via -r requirements.in
jsonpointer==3.0.0
    # via jsonschema
jsonschema[format]==4.25.1
    # via
    #   -r requirements.in
    #   hypothesis-jsonschema
    #   openapi-schema-validator
    #   openapi-spec-validator
    #   schemathesis
jsonschema-path==0.3.4
    # via openapi-spec-validator
jsonschema-specifications==2025.9.1
    # via
    #   jsonschema
    #   openapi-schema-validator
junit-xml==1.9
    # via schemathesis
kiwisolver==1.4.9
    # via matplotlib
lazy-object-proxy==1.12.0
    # via openapi-spec-validator
locust==2.42.6
    # via -r requirements.in
locust-cloud==1.29.5
    # via locust
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
matplotlib==3.10.7
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
msgpack==1.1.2
    # via locust
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via black
numpy==2.3.5
    # via
    #   contourpy
    #   matplotlib
    #   pandas
openapi-schema-validator==0.6.3
    # via openapi-spec-validator
openapi-spec-validator==0.7.2
    # via -r requirements.in
packaging==25.0
    # via
    #   black
    #   matplotlib
    #   prance
    #   pytest
pactman==2.31.0
    # via -r requirements.in
pandas==2.3.3
    # via -r requirements.in
pathable==0.4.4
    # via jsonschema-path
pathspec==0.12.1
    # via black
pillow==12.0.0
    # via matplotlib
platformdirs==4.5.1
    # via
    #   black
    #   locust-cloud
pluggy==1.6.0
    # via
    #   allure-python-commons
    #   pytest
    #   pytest-cov
ply==3.11
    # via jsonpath-ng
prance==25.4.8.0
    # via -r requirements.in
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
psutil==7.1.3
    # via locust
py-cpuinfo==9.0.0
    # via pytest-benchmark
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyparsing==3.2.5
    # via matplotlib
pyrate-limiter==3.9.0
    # via schemathesis
pytest==9.0.2
    # via
    #   -r requirements.in
    #   allure-pytest
    #   locust
    #   pactman
    #   pytest-asyncio
    #   pytest-benchmark
    #   pytest-cov
    #   pytest-html
    #   pytest-metadata
    #   pytest-mock
    #   pytest-subtests
    #   pytest-timeout
    #   pytest-xdist
    #   schemathesis
pytest-asyncio==1.3.0
    # via -r requirements.in
pytest-benchmark==5.2.3
    # via -r requirements.in
pytest-cov==7.0.0
    # via -r requirements.in
pytest-html==4.1.1
    # via -r requirements.in
pytest-metadata==3.1.1
    # via pytest-html
pytest-mock==3.15.1
    # via -r requirements.in
pytest-subtests==0.15.0
    # via schemathesis
pytest-timeout==2.4.0
    # via -r requirements.in
pytest-xdist==3.8.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   freezegun
    #   matplotlib
    #   pandas
python-dotenv==1.2.1
    # via -r requirements.in
python-engineio==4.12.3
    # via
    #   locust
    #   locust-cloud
    #   python-socketio
python-socketio[client]==5.15.0
    # via
    #   locust
    #   locust-cloud
pytokens==0.3.0
    # via black
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   jsonschema-path
    #   responses
    #   schemathesis
pyzmq==27.1.0
    # via locust
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   -r requirements.in
    #   jsonschema-path
    #   locust
    #   pactman
    #   prance
    #   python-socketio
    #   responses
    #   restnavigator
    #   schemathesis
    #   starlette-testclient
responses==0.25.8
    # via -r requirements.in
restnavigator==1.0.1
    # via pactman
rfc3339-validator==0.1.4
    # via
    #   jsonschema
    #   openapi-schema-validator
rfc3987==1.3.8
    # via jsonschema
rich==14.2.0
    # via schemathesis
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.16
    # via prance
ruamel-yaml-clib==0.2.15
    # via ruamel-yaml
schemathesis==4.7.4
    # via -r requirements.in
semver==3.0.4
    # via pactman
simple-websocket==1.1.0
    # via python-engineio
six==1.17.0
    # via
    #   junit-xml
    #   python-dateutil
    #   restnavigator
    #   rfc3339-validator
sortedcontainers==2.4.0
    # via hypothesis
starlette==0.50.0
    # via starlette-testclient
starlette-testclient==0.4.1
    # via schemathesis
tenacity==9.1.2
    # via
    #   -r requirements.in
    #   schemathesis
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   pytest-asyncio
    #   referencing
    #   schemathesis
    #   starlette
tzdata==2025.2
    # via
    #   arrow
    #   faker
    #   pandas
unidecode==1.4.0
    # via restnavigator
uri-template==1.3.0
    # via jsonschema
uritemplate==4.2.0
    # via restnavigator
urllib3==2.6.1
    # via
    #   geventhttpclient
    #   requests
    #   responses
webcolors==25.10.0
    # via jsonschema
websocket-client==1.9.0
    # via python-socketio
werkzeug==3.1.4
    # via
    #   flask
    #   flask-cors
    #   flask-login
    #   locust
    #   schemathesis
wsproto==1.3.2
    # via simple-websocket
yarl==1.22.0
    # via aiohttp
zope-event==6.1
    # via gevent
zope-interface==8.1.1
    # via gevent
